requirements.txt
flask>=3.0
flask-cors>=4.0
openpyxl>=3.1
Pillow>=10.0
psycopg2-binary>=2.9
